requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Storage Clients
qdrant-client==1.7.0
neo4j==5.14.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# NLP and ML
sentence-transformers==2.2.2
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
tiktoken==0.5.1
rank-bm25==0.2.2

# Document Processing
python-docx==1.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
scikit-learn==1.3.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
